requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
coverage==7.6.9
distro==1.9.0
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-modeltranslation==0.19.11
django-quill-editor==0.1.42
exceptiongroup==1.2.2
execnet==2.1.1
Faker==33.1.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
openai==1.57.2
packaging==24.2
pillow==11.0.0
pip-autoremove==0.10.0
pip-chill==1.0.3
pipdeptree==2.24.0
pluggy==1.5.0
promise==2.3
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
text-unidecode==1.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
